tests now work with the new setting up of directories.
